Fix deadlock when no watchers can be instrumented
This implements a simple message passing scheme.  We don't do anything
to ensure linearized read/write of the underlying semaphore, since it's
guarded behind a futex which is presumably a heavier-handed coordination
primitive in the first place.

I still need to research what happens when a file descriptor is closed
after being sent over a USD, but before it gets fully received.  My gut
instinct is that it is OK, but we proceed safely for now.

I simplified some of the early-exit logic by removing it.  We now
iterate through ALL possible CPUs and ALL possible watchers rather than
trying to be clever.   This is handled in poll() since it ignores
negative file descriptors anyway.