Add some more errors to CI; suppress optimization
* GCC and Clang both use a clever, but aggressive, lookup-based
optimization strategy to collapse call stacks in collatz.  We don't want
that, since collatz needs to generate big, lovely stacks for use in
benchmarking.  Teaching the harness to filter out the -O2 which has been
set for recent non-debug builds.  Also made sure `-g` was added in
manually, just in case.

* The Collatz build had bit-rotted since it's a nonoptional, non-deployed
product.  Fixed CI and made the build non-optional so this doesn't rot
out.

Unrelated, but maybe it's time to check on S3 aging policies on our
repo.  :)

* Fix unbound variable for promotion