Protect instrumented process from spurious SIGCHLD
The typical mistake first-timers make when writing multiprocessing code
is assuming code sequentiality implies process sequence.  I'm supposed
to know better by now, but I continue to make this particular blunder
with regularity.

The latest such example is the realization that the proces spawned by
the first `fork()` in the instrumentation sequence may very well hang
around until well after `execve()` is called in the main process.  This
scenario is discussed in the code comments, but the short story is that
many services will break or segfault if you do this to them.

I believe I have a correct and simple yield pattern here, waiting for
the child to get cleaned up (specifically:  waiting for the SIGCHLD to
get consumed!) without concern for hanging.