Move dwfl lifetime into ddprof's export cycle
A previous PR changed the cadence of dwfl updates from once to
once-per-sample.  This actually has a measurable, negative impact on
performance, even though it controlled certain aspects of memory
allocation.  We're testing out a middle-ground strategy where the dwfl
is cleared every export.  This amortizes infrequent operations along the
duration of serializing and sending an HTTP request, which seems like a
modestly long timescale anyway.  :)