Allow intermediate failures to disengage callback
Some of the intermediate operations performed by the profiler are
failable.  We've always allowed the profiler to continue in a degraded
state (presumably it would recover or get killed some other way), but
since we're trying to be more strict about things, it's now necessary
for the callbacks consumed by perf.c main_loop() to break the loop.

This patch refactors those callbacks to have boolean return and causes
the main_loop to check return and exit on failure.  It adds two such
failure modes related to the release of intermediate caches.