Call main_loop:finish on every worker exit path
In the previous commit, we introduced the `finish_fun()` callback, which
only executed when the worker process was about to die *and* it was
signalling that no further workers should be spawned.  However, there is
valid cleanup which could occur at the termination of every worker.

It was possible to have two callbacks, but instead we just added a
boolean to the callback parameter list, specifying whether or not this
is a terminal close for profiling or merely a worker refresh.