Fix cppcheck issues, remove niladic macro expansion
This patch fixes a few issues caught by cppcheck, including some of the
suppressed ones.  Such suppressed checks were removed from the file.

Adds the ddprof directory to cppcheck, since the interface relies on a
few macros.

Move some single-use macro definitions out of ddprof.h into main.c

The biggest change in this patch is to the behavior of ddres and the
logger.  The logger will now take a null format string as a suppression
of logging.  There is an associated change to ddres.h which removes the
use of `__VA_OPT__` in favor of passing down this special value to
suppress logging, with appropriate changes in the codebase when a
niladic value for `__VA_ARGS__` was otherwise passed.  The reason for
this change is primarily to conform with the level of support for
`__VA_OPT__` in cppcheck and other tools I checked.  There's every
reason to believe it's going to become a part of the next standard, and
arguably its absence has been a breathtakingly impactful omission in the
standard since the inception of variadic macros, but it *seems* that we
might be able to do without, and doing so enhances the coverage of our
static analysis tools, which are otherwised forced to omit expansion of
any macro which, during preprocessing, consumes `__VA_OPT__`.