Remove the magic number 3
The storage container for x86 registers is an array of 64-bit integers.
It's necessary to know the size of this array in places, which is
specified in metadata nowhere as far as I know as the size is inferred
from the original input to `perf_event_open()`.  In places, we hardcoded
the value 3, since that's how many bits we had set in the mask.

This patch better documents the intent behind the number 3 by replacing
it with the `PERF_REGS_COUNT` macro (which at least is easier to grep
than 3)