Remove export logic from worker_finish
Previously, we had a send_final parameter which could be used to ensure
the last partial export before shutdown was sent to the backend.  Some
of the additional logic we'd need to reintroduce this feature with
fork-workers is a little hairy, so rather than keeping it in vestigial
form, I'm removing it altogether.