Add simple lookup to collrunner
Collrunner has been broken for a while, since the way we're building the
project is different now.  This fixes collrunner to use the user's
environment settings to locate the binary.  This presumes the user has
build ddprof and the benchmarking program(s) in the same place, which
should generally be true.