Filter out pid 0
PID 0 appears to be related to runc or other aspects of the container
management system.  Nothing about these PIDs can be utilized and in many
systems they cause the native profiler to spend an outrageous amount of
overhead on nothing useful.

Eventually these will probably be pushed into the profile, which can be
done by modifying the guard for the sample-type event.