Slight changes to ringbuffer behavior
Previously, the ringbuffer would only be drained if the underlying file
descriptor was hot.  This was slightly problematic, since it meant we
had to wait for a kernel->userspace message to go through all the
requisite machinery when we could just check the ringbuffer and call it
good.  This patch changes the behavior of the ringbuffer draining.

1.  Any time poll() activates in a non-erroneous (e.g., all instrumented
    processes die) state, check *all* ringbuffers
2.  If nothing is going on, wake up every 10ms
3.  Do not check for a 0 return as the only means of assessing whether a
    timeout call is necessary.  Instead, on every pass check whether any
    of the events were samples, and if they were not, then call the
    timeout code.

In a future patch, we'll reorder the events.

This patch massively decreases (1000->10) the period of idle activation