0.6.4 - Erwan.viollet/v frames (#142)
- Virtual frames
Check information from DSO when we are not able to get source information from dwarf
Insert frames when we are not able to unwind. Including:
    Truncated stacks
    Unable to find DSOs
    Unwinding errors (use info from DSO)

- Split Dwfl info per process
Associate Dwfl objects to the PID of the process we are unwinding.

- Native bench
Add a benchmark around the sleep script to test overhead of applications that fork a lot.