Reduce memory overhead of symbolization cache (#148)
- Use a unique ID for DSOs to identify when symbols can be reused
Consider the normalised address to ensure the access to symbols is coherent for different processes (with randomized addresses)
This is very important for processes that fork a lot. The current implementation will quickly use all memory available.

- Use a tree structure instead of a hashmap to store symbols
Read the address range from the elf information extracted from dwarf library.
Tests : I validated that there are no errors on a few examples. I still have a mechanism that validates the accuracy of the cache (through environment variable)
This is the change I am least certain about. I will change it back if we seen any kind of regression.

- Fix the mapping information sent through the exporter 
We should still review what I am sending (I think the info is still not very relevant)

- Standardize lookup APIs

- Fix usage of GDB in run.sh