Dwarf lookup cache (#155)
- Symbol cache
Fix the behaviour with ranges taking into account bias to compute the address into the mapped region.
Add a unit test to capture the logic behind the pointer adjustments.

- Leak fix
Minor leak fix when we fail to mmap.

- Fallback mechanism: Switch to a hash-map implementation
Protected by a #define the symbol cache can use a hash-map, this allows a fallback mechanism if any issues occur. It also allows us to compare the efficiency in lookups.