Sanchda/clang (#201)
Add support for clang builds and makes it the default.  Primarily, this is implemented by adding a new helper function to the cmake (via cmake/Helperfunc.cmake) and modifying build dependencies.  This helper function allows the caller to specify whether a given target is dynamically or statically built against LLVM’s libc++—although it is generally called statically as the test context does not presume the same dependencies as the build context.

Note that this implementation is a little tricky, as it passes a somewhat bizarre set of compiler and linker flags (sometimes duplicates of each other) to the driver.

Updates elfutils to 186, which is the minimum required for clang support
Quiet several warnings in the elfutils build in order to make the build pass, despite nominal clang support in elfutils 186